requirements.txt
python-telegram-bot==21.6
openai==0.28.0
python-dotenv==1.0.1